Engine: refactor Stream hierarchy into a Stream - StreamBase pair
Old state:
- IStream is a pure virtual interface for the basic stream class.
- Stream is a base class for all streams, every stream type is derived from it, or from DataStream specifically.
- Stream interface mixes both low-level operations and helper methods, e.g. ones that read and write independent of system endianess.
- IStream is matched by the engine's plugin API (IAGSStream).

New state:
- IStreamBase is a pure virtual interface for the basic stream class.
- StreamBase is a parent class for derived stream "device" implementations (File, Memory, etc).
- Stream is a thin wrapper over a IStreamBase pointer with no virtual members.
- Stream owns IStreamBase object, unless ordered to give it away it by calling ReleaseStreamBase().
- DataStream class is removed (because Stream class now has the necessary endianess conversions).
- IStreamBase is matched by the engine's plugin API (IAGSStream).

IStreamBase provides basic operations over a stream.
Stream works with IStreamBase and provides a multitude of helper methods, similar to what DataStream had been before.
Stream owns IStreamBase and disposes it on destruction by default, but it's possible to release StreamBase's ownership with a call to ReleaseStreamBase(). This ability is left for cases when you need to wrap StreamBase temporarily, but keep it after use.

Besides changes to the stream classes, changes to the rest of the engine code are:
- Constructing a Stream object now requires to pass a IStreamBase ptr, similar to:

    Stream s(std::make_unique<FileStream>(...args...));

IMPORTANT NOTE:
Stream class is used everywhere, and is left as-is for now. In theory, some things could instead use bare IStreamBase. AssetManager and File helpers could return IStreamBase instead. But this is postponed for the future consideration. Currently there will be only one place where we would have to recover IStreamBase out of the Stream, and that's when the stream is given to the plugin cast to IAGSStream.